--- a/Cargo.toml
+++ b/Cargo.toml
@@ -15,6 +15,7 @@ composability, and iterator-like interfaces.
 
 [dependencies]
 log = { version = "0.3", default-features = false }
+slab = "0.3.0"
 
 [features]
 use_std = []