--- a/src/stream/buffer_unordered.rs
+++ b/src/stream/buffer_unordered.rs
@@ -1,7 +1,12 @@
 use std::prelude::v1::*;
+use std::sync::Arc;
+use std::collections::VecDeque;
+
+use task::{self, EventSet, UnparkEvent};
 
 use {Async, IntoFuture, Poll, Future};
 use stream::{Stream, Fuse};
+use stack::Stack;
 
 /// An adaptor for a stream of futures to execute the futures concurrently, if
 /// possible, delivering results as they become available.
@@ -17,6 +22,7 @@ pub struct BufferUnordered<S>
     stream: Fuse<S>,
     futures: Vec<Option<<S::Item as IntoFuture>::Future>>,
     cur: usize,
+    stack: Arc<Stack<usize>>,
 }
 
 pub fn new<S>(s: S, amt: usize) -> BufferUnordered<S>
@@ -27,6 +33,7 @@ pub fn new<S>(s: S, amt: usize) -> BufferUnordered<S>
         stream: super::fuse::new(s),
         futures: (0..amt).map(|_| None).collect(),
         cur: 0,
+        stack: Arc::new(Stack::new()),
     }
 }
 
@@ -39,11 +46,12 @@ impl<S> Stream for BufferUnordered<S>
 
     fn poll(&mut self) -> Poll<Option<Self::Item>, Self::Error> {
         // First, try to fill in all the futures
-        for future in &mut self.futures {
+        for (idx, future) in self.futures.iter_mut().enumerate() {
             if future.is_none() {
                 match try!(self.stream.poll()) {
                     Async::Ready(Some(s)) => {
                         *future = Some(s.into_future());
+                        self.stack.push(idx);
                     }
                     Async::Ready(None) => break,
                     Async::NotReady => break,
@@ -54,25 +62,31 @@ impl<S> Stream for BufferUnordered<S>
         // Next, try and step futures forward until we find a ready one.
         // Always start at `cur` for fairness.
         let mut waiting = false;
-        for i in 0..self.futures.len() {
-            let mut idx = self.cur + i;
-            if idx >= self.futures.len() {
-                idx -= self.futures.len();
-            }
+        let mut drained = self.stack.drain();
+        while let Some(idx) = drained.next() {
             let future = &mut self.futures[idx];
             let result = match *future {
-                Some(ref mut s) => match s.poll() {
-                    Ok(Async::NotReady) => {
-                        waiting = true;
-                        continue
-                    },
-                    Ok(Async::Ready(e)) => Ok(Async::Ready(Some(e))),
-                    Err(e) => Err(e),
+                Some(ref mut s) => {
+                    let event = UnparkEvent::new(self.stack.clone(), idx);
+                    let poll_result = task::with_unpark_event(event, || {
+                        s.poll()
+                    });
+                    match poll_result {
+                        Ok(Async::NotReady) => {
+                            waiting = true;
+                            continue
+                        },
+                        Ok(Async::Ready(e)) => Ok(Async::Ready(Some(e))),
+                        Err(e) => Err(e),
+                    }
                 },
                 None => continue,
             };
-            self.cur = i + 1;
             *future = None;
+            // push unresolved futures back
+            for idx in drained {
+                self.stack.push(idx);
+            }
             return result;
         }