--- a/src/stream/buffer_unordered.rs
+++ b/src/stream/buffer_unordered.rs
@@ -2,6 +2,7 @@ use std::prelude::v1::*;
 use std::sync::Arc;
 use std::collections::VecDeque;
 
+use slab::Slab;
 use task::{self, EventSet, UnparkEvent};
 
 use {Async, IntoFuture, Poll, Future};
@@ -20,7 +21,7 @@ pub struct BufferUnordered<S>
           S::Item: IntoFuture,
 {
     stream: Fuse<S>,
-    futures: Vec<Option<<S::Item as IntoFuture>::Future>>,
+    futures: Slab<<S::Item as IntoFuture>::Future>,
     cur: usize,
     stack: Arc<Stack<usize>>,
 }
@@ -31,7 +32,7 @@ pub fn new<S>(s: S, amt: usize) -> BufferUnordered<S>
 {
     BufferUnordered {
         stream: super::fuse::new(s),
-        futures: (0..amt).map(|_| None).collect(),
+        futures: Slab::with_capacity(amt),
         cur: 0,
         stack: Arc::new(Stack::new()),
     }
@@ -46,16 +47,18 @@ impl<S> Stream for BufferUnordered<S>
 
     fn poll(&mut self) -> Poll<Option<Self::Item>, Self::Error> {
         // First, try to fill in all the futures
-        for (idx, future) in self.futures.iter_mut().enumerate() {
-            if future.is_none() {
-                match try!(self.stream.poll()) {
-                    Async::Ready(Some(s)) => {
-                        *future = Some(s.into_future());
-                        self.stack.push(idx);
-                    }
-                    Async::Ready(None) => break,
-                    Async::NotReady => break,
+        while self.futures.len() < self.futures.capacity() {
+            match try!(self.stream.poll()) {
+                Async::Ready(Some(s)) => {
+                    let idx = match self.futures.insert(s.into_future()) {
+                        Ok(idx) => idx,
+                        // Can't unwrap because future does not implement Debug
+                        Err(_original_future) => unreachable!(),
+                    };
+                    self.stack.push(idx);
                 }
+                Async::Ready(None) => break,
+                Async::NotReady => break,
             }
         }
 
@@ -64,8 +67,7 @@ impl<S> Stream for BufferUnordered<S>
         let mut waiting = false;
         let mut drained = self.stack.drain();
         while let Some(idx) = drained.next() {
-            let future = &mut self.futures[idx];
-            let result = match *future {
+            let result = match self.futures.get_mut(idx) {
                 Some(ref mut s) => {
                     let event = UnparkEvent::new(self.stack.clone(), idx);
                     let poll_result = task::with_unpark_event(event, || {
@@ -82,7 +84,7 @@ impl<S> Stream for BufferUnordered<S>
                 },
                 None => continue,
             };
-            *future = None;
+            self.futures.remove(idx);
             // push unresolved futures back
             for idx in drained {
                 self.stack.push(idx);