This is an alternative to https://github.com/tokio-rs/tokio-core/pull/60

While this is much faster, I'm not sure it's as scalable, here are some benchmarks.

So I'm testing a [redis server implementation](https://github.com/tailhook/tokio-tutorial/blob/05acb04df8e155312f6b13ffde6896a79a97f84a/redis_bu/src/main.rs#L61-L65).

The method is the following. I'm running a bunch of redis-benchmark processes with `-I` flag:

```
redis-benchmark -c 20000 -I -p 7001 -h 127.0.0.$i
```

It creates inactive connections (we need to use `127.0.0.x` IPs to avoid the limit for ephemeral ports) 

And then benchmark GET,SET on usually 50 connections:

```
redis-benchmark -c 1000 -p 7001 -n 100000 GET,SET
```
## Benchmark Results

| Test | RPS | Idle | Active |
| --- | --- | --- | --- |
| This patch | 129836.41 | 0 | 50 |
| This patch | 62770.70 | 40k | 50 |
| This patch | 46593.98 | 60k | 50 |
| This patch | 36387.45 | 80k | 50 |
| original | 3418.80 | 0 | 50 |
| original | 42698.55 | 0 | 1000 |
| original | 2477.70 | 20k | 50 |
| original | 36697.25 | 20k | 1000 |
| tokio-core | 133280.03 | 0 | 50 |
| tokio-core | 127877.23 | 80k | 50 |

As you can see with the patch applied performance matches the performance of tokio core. Still it doesn't scale well with increasing the number of idle connections.

Other interesting observations is that before this patch performance increases with the number of active connections (well, It's understandable in the retrospective).
## Implementation

While the implementation is far from being perfect, it passes the tests and the only inefficiency I can see is [here](https://github.com/alexcrichton/futures-rs/compare/master...tailhook:fast_buffer_unordered?expand=1#diff-236773cc553adb1ea6bfcc851419c09dR88). But it would only influence benchmarks if they were closing connections often. These benchmarks keep keep-alive connection for the whole duration of the test.

Another issue is that `buffer_unordered` is probably not very good combinator for this use case because we don't want to return any result at all (and non-returning a value allows us to fix the bug above more efficiently). So should it be something like `consume()` ?