@tailhook 2016-10-28
Well, I'm feeling stupid (my excuse is that there were no coffee at hack place :) )

Added `Slab` to the `buffer_unordered` and got `141k - 146k` on the same machine with difference being not significant in all tests.

Basically, its now faster that using `spawn` from tokio for some reason (I have few guesses, but..)

So questions left are:

1. What to do with slab? Can we depend on it? Should this combinator be in a separate crate? Should we embed a stripped down implementation of a slab?
2. Should we have a `consume` variant (without a return value)?
3. We need Box for `stack`. How can we use it in stable rust with `no_std`? (I'm not good with `no_std` so maybe I'm missing something, but `alloc` crate, which originally contains box, is unstable)