@tailhook 2016-10-28
And the memory usage:

|Mem (kb) | Conns    |                                 |
|---------|----------|---------------------------------|
|403124   | 20k      | Buffer unordered                |
|318536   | 20k -> 0 | After all connections are closed|
|500612   | 40k      | Buffer unordered                |
|122144   | 20k      | Spawned                         |
| 23180   | 20k -> 0 | After all connections are closed|
|223548   | 40k      | Spawned                         |

While overall memory usage is disappointing (~5k per idle connection), this patch doesn't change the per-connection memory usage too much (I need more tests to measure real difference). The constant part + 300Mb is probably because of a slab for 1M connection we have configured in example.