@alexcrichton 2016-10-28
> Basically, its now faster that using spawn from tokio for some reason (I have few guesses, but..)

Whoa! That's unexpected... I wonder if it's because there's only one virtual dispatch (to the one stream) vs N virtual dispatches (one per stream). In any case, promising numbers!

> What to do with slab? Can we depend on it? Should this combinator be in a separate crate? Should we embed a stripped down implementation of a slab?

Let's vendor a stripped down version for now and see how far we can get with that.

> Should we have a consume variant (without a return value)?

Hm I'm not sure I follow, can you remind me what this is?

> We need Box for stack. How can we use it in stable rust with no_std? (I'm not good with no_std so maybe I'm missing something, but alloc crate, which originally contains box, is unstable)

Ah many combinators just aren't available in `no_std`, so this combinator can just be turned off in the `no_std` case.

---

Also I had an idea for dealing with leftover nodes in the stack that we drained if a stream resolves. We should just have `pending: Option<Drain<T>>` on the stream itself. That way we never need to re-enqueue and it should remain just as efficient in theory!