@alexcrichton 2016-11-05
Ok! I've cleaned up a few pieces here and there and otherwise landed this now with small vendorings of `Stack` and `Slab`. I added a test as well and covered a few other use cases I believe. Let me know if it doesn't work out though!